EICNET-1481: Create an abstract class to be used by all group related resources.
